Camera Update
Camera now follows a point further up the road instead of an offset right in front of the car. This prevents the camera pointing to an empty spot in front of a corner instead of focusing on a point around the corner. This offset also sllightly increases by the current speed of the car. Still tweaking the values to find the right middle-ground between the different values that are used to calculate the offset (distance ahead on the road, speed of the camera). Too much offset is nauseating in early testing, and if the followspeed is too slow the camera lags behind the player, which is the oppsite of what we want..